Core: Updated to 12.0.7
Port From (https://github.com/TrinityCore/TrinityCore/commit/c8ec3740265a237bc4da2ded4641203ae4fc102d)

+++ Source/Game/Entities/Item/ItemTemplate.cs
@@ -322,7 +322,7 @@ namespace Game.Entities
         public uint GetSellPrice() { return ExtendedData.SellPrice; }
         public InventoryType GetInventoryType() { return ExtendedData.inventoryType; }
         public int GetAllowableClass() { return ExtendedData.AllowableClass; }
-        public RaceMask<long> GetAllowableRace() { return ExtendedData.AllowableRace; }
+        public RaceMask<int> GetAllowableRace() { return ExtendedData.AllowableRace; }
         public uint GetBaseItemLevel() { return ExtendedData.ItemLevel; }
         public int GetBaseRequiredLevel() { return ExtendedData.RequiredLevel; }
         public uint GetRequiredSkill() { return ExtendedData.RequiredSkill; }
